[modify] PhotoListAdapter - load image based on dimen
In the previous code, the image was loaded by using imageView as an
argument. However, when calculating the size of the image, it will lead
to width and height of 0, causing the app to crash.

In order to solve this problem, the itemholder width and height are
passed into the function instead, knowing that the imageView is the same
height and width as the itemView.

This is done by calling holder.itemView.post to get the proper dimension
of the itemView.